// src/shared/types/users.ts
export interface User {
  apellido:  string;
  documento: string;
  cliente:   string;
  entidad:   string;
  codigo:    string;
  fec_mae:   string;
  cuopag:    string;
  cuopen:    string;
  ult_cob:   string;
  fupdate:   string;
  fupd:      string;
  ganador:   string;
  fec_gan:   string;
  pre_pen:   string;
}